moved functions from passport to auth, they fit there better
